Refactor extract conversation FileWidget
The FileWidget is only ever used as part of a conversation,
and the new package structure reflects that.

I only touched the related files and I grouped them around
domain concepts and features.
