+++ securedrop_client/gui/widgets.py
@@ -26,7 +26,7 @@ from uuid import uuid4
 
 import arrow
 import sqlalchemy.orm.exc
-from PyQt5.QtCore import QEvent, QObject, QSize, Qt, QTimer, pyqtBoundSignal, pyqtSignal, pyqtSlot
+from PyQt5.QtCore import QEvent, QObject, QSize, Qt, QTimer, pyqtSignal, pyqtSlot
 from PyQt5.QtGui import (
     QBrush,
     QColor,
@@ -34,24 +34,18 @@ from PyQt5.QtGui import (
     QFocusEvent,
     QFont,
     QIcon,
-    QKeyEvent,
     QKeySequence,
     QLinearGradient,
     QMouseEvent,
     QPalette,
-    QPixmap,
     QResizeEvent,
 )
 from PyQt5.QtWidgets import (
     QAction,
-    QApplication,
-    QDialog,
-    QDialogButtonBox,
     QGraphicsDropShadowEffect,
     QGridLayout,
     QHBoxLayout,
     QLabel,
-    QLineEdit,
     QListWidget,
     QListWidgetItem,
     QMenu,
@@ -75,7 +69,6 @@ from securedrop_client.db import (
     Source,
     User,
 )
-from securedrop_client.export import ExportError, ExportStatus
 from securedrop_client.gui import (
     SDPushButton,
     SecureQLabel,
@@ -83,10 +76,11 @@ from securedrop_client.gui import (
     SvgPushButton,
     SvgToggleButton,
 )
+from securedrop_client.gui.conversation import FileWidget  # should eventually only be Conversation
+from securedrop_client.gui.dialogs import ModalDialog  # should eventually come from gui
 from securedrop_client.logic import Controller
 from securedrop_client.resources import load_css, load_icon, load_image, load_movie
 from securedrop_client.storage import source_exists
-from securedrop_client.utils import humanize_filesize
 
 logger = logging.getLogger(__name__)
 
@@ -1715,34 +1709,6 @@ class LoginErrorBar(QWidget):
         self.hide()
 
 
-class PasswordEdit(QLineEdit):
-    """
-    A LineEdit with icons to show/hide password entries
-    """
-
-    def __init__(self, parent: QDialog) -> None:
-        self.parent = parent
-        super().__init__(self.parent)
-
-        self.visibleIcon = load_icon("eye_visible.svg")
-        self.hiddenIcon = load_icon("eye_hidden.svg")
-
-        self.setEchoMode(QLineEdit.Password)
-        self.togglepasswordAction = self.addAction(self.hiddenIcon, QLineEdit.TrailingPosition)
-        self.togglepasswordAction.triggered.connect(self.on_toggle_password_Action)
-        self.password_shown = False
-
-    def on_toggle_password_Action(self) -> None:
-        if not self.password_shown:
-            self.setEchoMode(QLineEdit.Normal)
-            self.password_shown = True
-            self.togglepasswordAction.setIcon(self.visibleIcon)
-        else:
-            self.setEchoMode(QLineEdit.Password)
-            self.password_shown = False
-            self.togglepasswordAction.setIcon(self.hiddenIcon)
-
-
 class SenderIcon(QWidget):
     """
     Represents a reply to a source.
@@ -2167,848 +2133,6 @@ class ReplyWidget(SpeechBubble):
         self.color_bar.setStyleSheet("")
         self.color_bar.setObjectName("ReplyWidget_status_bar_failed")
         self.color_bar.setStyleSheet(self.STATUS_BAR_CSS)
-
-
-class FileWidget(QWidget):
-    """
-    Represents a file.
-    """
-
-    DOWNLOAD_BUTTON_CSS = load_css("file_download_button.css")
-
-    TOP_MARGIN = 18
-    BOTTOM_MARGIN = 0
-    FILE_FONT_SPACING = 2
-    FILE_OPTIONS_FONT_SPACING = 1.6
-    FILENAME_WIDTH_PX = 360
-    FILE_OPTIONS_LAYOUT_SPACING = 8
-
-    WIDTH_TO_CONTAINER_WIDTH_RATIO = 5 / 9
-    MIN_CONTAINER_WIDTH = 750
-    MIN_WIDTH = 400
-
-    def __init__(
-        self,
-        file_uuid: str,
-        controller: Controller,
-        file_ready_signal: pyqtBoundSignal,
-        file_missing: pyqtBoundSignal,
-        index: int,
-        container_width: int,
-    ) -> None:
-        """
-        Given some text and a reference to the controller, make something to display a file.
-        """
-        super().__init__()
-
-        self.controller = controller
-        self.file = self.controller.get_file(file_uuid)
-        self.uuid = file_uuid
-        self.index = index
-        self.downloading = False
-
-        self.adjust_width(container_width)
-
-        self.setObjectName("FileWidget")
-        file_description_font = QFont()
-        file_description_font.setLetterSpacing(QFont.AbsoluteSpacing, self.FILE_FONT_SPACING)
-        self.file_buttons_font = QFont()
-        self.file_buttons_font.setLetterSpacing(
-            QFont.AbsoluteSpacing, self.FILE_OPTIONS_FONT_SPACING
-        )
-
-        # Set layout
-        layout = QHBoxLayout()
-        self.setLayout(layout)
-
-        # Set margins and spacing
-        layout.setContentsMargins(0, self.TOP_MARGIN, 0, self.BOTTOM_MARGIN)
-        layout.setSpacing(0)
-
-        # File options: download, export, print
-        self.file_options = QWidget()
-        self.file_options.setObjectName("FileWidget_file_options")
-        file_options_layout = QHBoxLayout()
-        self.file_options.setLayout(file_options_layout)
-        file_options_layout.setContentsMargins(0, 0, 0, 0)
-        file_options_layout.setSpacing(self.FILE_OPTIONS_LAYOUT_SPACING)
-        file_options_layout.setAlignment(Qt.AlignLeft)
-        self.download_button = QPushButton(_(" DOWNLOAD"))
-        self.download_button.setObjectName("FileWidget_download_button")
-        self.download_button.setStyleSheet(self.DOWNLOAD_BUTTON_CSS)
-        self.download_button.setSizePolicy(QSizePolicy.Fixed, QSizePolicy.Fixed)
-        self.download_button.setIcon(load_icon("download_file.svg"))
-        self.download_button.setFont(self.file_buttons_font)
-        self.download_button.setCursor(QCursor(Qt.PointingHandCursor))
-        self.download_animation = load_movie("download_file.gif")
-        self.export_button = QPushButton(_("EXPORT"))
-        self.export_button.setObjectName("FileWidget_export_print")
-        self.export_button.setFont(self.file_buttons_font)
-        self.export_button.setCursor(QCursor(Qt.PointingHandCursor))
-        self.middot = QLabel("·")  # nosemgrep: semgrep.untranslated-gui-string
-        self.print_button = QPushButton(_("PRINT"))
-        self.print_button.setObjectName("FileWidget_export_print")
-        self.print_button.setFont(self.file_buttons_font)
-        self.print_button.setCursor(QCursor(Qt.PointingHandCursor))
-        file_options_layout.addWidget(self.download_button)
-        file_options_layout.addWidget(self.export_button)
-        file_options_layout.addWidget(self.middot)
-        file_options_layout.addWidget(self.print_button)
-
-        self.download_button.installEventFilter(self)
-        self.export_button.clicked.connect(self._on_export_clicked)
-        self.print_button.clicked.connect(self._on_print_clicked)
-
-        self.file_name = SecureQLabel(
-            wordwrap=False, max_length=self.FILENAME_WIDTH_PX, with_tooltip=True
-        )
-        self.file_name.setObjectName("FileWidget_file_name")
-        self.file_name.installEventFilter(self)
-        self.file_name.setCursor(QCursor(Qt.PointingHandCursor))
-        self.no_file_name = SecureQLabel(_("ENCRYPTED FILE ON SERVER"), wordwrap=False)
-        self.no_file_name.setObjectName("FileWidget_no_file_name")
-        self.no_file_name.setFont(file_description_font)
-
-        # Line between file name and file size
-        self.horizontal_line = QWidget()
-        self.horizontal_line.setObjectName("FileWidget_horizontal_line")
-        self.horizontal_line.setSizePolicy(QSizePolicy.Expanding, QSizePolicy.Fixed)
-
-        # Space between elided file name and file size when horizontal line is hidden
-        self.spacer = QWidget()
-        self.spacer.setSizePolicy(QSizePolicy.Expanding, QSizePolicy.Fixed)
-        self.spacer.hide()
-
-        # File size
-        self.file_size = SecureQLabel(humanize_filesize(self.file.size))
-        self.file_size.setObjectName("FileWidget_file_size")
-        self.file_size.setAlignment(Qt.AlignRight)
-
-        # Decide what to show or hide based on whether or not the file's been downloaded
-        self._set_file_state()
-
-        # Add widgets
-        layout.addWidget(self.file_options)
-        layout.addWidget(self.file_name)
-        layout.addWidget(self.no_file_name)
-        layout.addWidget(self.spacer)
-        layout.addWidget(self.horizontal_line)
-        layout.addWidget(self.file_size)
-
-        # Connect signals to slots
-        file_ready_signal.connect(self._on_file_downloaded, type=Qt.QueuedConnection)
-        file_missing.connect(self._on_file_missing, type=Qt.QueuedConnection)
-
-    def adjust_width(self, container_width: int) -> None:
-        """
-        This is a workaround to the workaround for https://bugreports.qt.io/browse/QTBUG-85498.
-        See comment in the adjust_width method for SpeechBubble.
-        """
-        if container_width < self.MIN_CONTAINER_WIDTH:
-            self.setFixedWidth(self.MIN_WIDTH)
-        else:
-            self.setFixedWidth(container_width * self.WIDTH_TO_CONTAINER_WIDTH_RATIO)
-
-    def eventFilter(self, obj: QObject, event: QEvent) -> None:
-        t = event.type()
-        if t == QEvent.MouseButtonPress:
-            if event.button() == Qt.LeftButton:
-                self._on_left_click()
-        elif t == QEvent.HoverEnter and not self.downloading:
-            self.download_button.setIcon(load_icon("download_file_hover.svg"))
-        elif t == QEvent.HoverLeave and not self.downloading:
-            self.download_button.setIcon(load_icon("download_file.svg"))
-
-        return QObject.event(obj, event)
-
-    def update_file_size(self) -> None:
-        try:
-            self.file_size.setText(humanize_filesize(self.file.size))
-        except Exception as e:
-            logger.error(f"Could not update file size on FileWidget: {e}")
-            self.file_size.setText("")
-
-    def _set_file_state(self) -> None:
-        if self.file.is_decrypted:
-            logger.debug("Changing file {} state to decrypted/downloaded".format(self.uuid))
-            self._set_file_name()
-            self.download_button.hide()
-            self.no_file_name.hide()
-            self.export_button.show()
-            self.middot.show()
-            self.print_button.show()
-            self.file_name.show()
-            self.update_file_size()
-        else:
-            logger.debug("Changing file {} state to not downloaded".format(self.uuid))
-            self.download_button.setText(_("DOWNLOAD"))
-
-            # Ensure correct icon depending on mouse hover state.
-            if self.download_button.underMouse():
-                self.download_button.setIcon(load_icon("download_file_hover.svg"))
-            else:
-                self.download_button.setIcon(load_icon("download_file.svg"))
-
-            self.download_button.setFont(self.file_buttons_font)
-            self.download_button.show()
-
-            # Reset stylesheet
-            self.download_button.setStyleSheet("")
-            self.download_button.setObjectName("FileWidget_download_button")
-            self.download_button.setStyleSheet(self.DOWNLOAD_BUTTON_CSS)
-
-            self.no_file_name.hide()
-            self.export_button.hide()
-            self.middot.hide()
-            self.print_button.hide()
-            self.file_name.hide()
-            self.no_file_name.show()
-
-    def _set_file_name(self) -> None:
-        self.file_name.setText(self.file.filename)
-        if self.file_name.is_elided():
-            self.horizontal_line.hide()
-            self.spacer.show()
-
-    @pyqtSlot(str, str, str)
-    def _on_file_downloaded(self, source_uuid: str, file_uuid: str, filename: str) -> None:
-        if file_uuid == self.uuid:
-            self.downloading = False
-            QTimer.singleShot(300, self.stop_button_animation)
-
-    @pyqtSlot(str, str, str)
-    def _on_file_missing(self, source_uuid: str, file_uuid: str, filename: str) -> None:
-        if file_uuid == self.uuid:
-            self.downloading = False
-            QTimer.singleShot(300, self.stop_button_animation)
-
-    @pyqtSlot()
-    def _on_export_clicked(self) -> None:
-        """
-        Called when the export button is clicked.
-        """
-        if not self.controller.downloaded_file_exists(self.file):
-            return
-
-        self.export_dialog = ExportDialog(self.controller, self.uuid, self.file.filename)
-        self.export_dialog.show()
-
-    @pyqtSlot()
-    def _on_print_clicked(self) -> None:
-        """
-        Called when the print button is clicked.
-        """
-        if not self.controller.downloaded_file_exists(self.file):
-            return
-
-        dialog = PrintDialog(self.controller, self.uuid, self.file.filename)
-        dialog.exec()
-
-    def _on_left_click(self) -> None:
-        """
-        Handle a completed click via the program logic. The download state
-        of the file distinguishes which function in the logic layer to call.
-        """
-        # update state
-        self.file = self.controller.get_file(self.uuid)
-
-        if self.file.is_decrypted:
-            # Open the already downloaded and decrypted file.
-            self.controller.on_file_open(self.file)
-        elif not self.downloading:
-            if self.controller.api:
-                self.start_button_animation()
-            # Download the file.
-            self.controller.on_submission_download(File, self.uuid)
-
-    def start_button_animation(self) -> None:
-        """
-        Update the download button to the animated "downloading" state.
-        """
-        self.downloading = True
-        self.download_animation.frameChanged.connect(self.set_button_animation_frame)
-        self.download_animation.start()
-        self.download_button.setText(_(" DOWNLOADING "))
-
-        # Reset widget stylesheet
-        self.download_button.setStyleSheet("")
-        self.download_button.setObjectName("FileWidget_download_button_animating")
-        self.download_button.setStyleSheet(self.DOWNLOAD_BUTTON_CSS)
-
-    def set_button_animation_frame(self, frame_number: int) -> None:
-        """
-        Sets the download button's icon to the current frame of the spinner
-        animation.
-        """
-        self.download_button.setIcon(QIcon(self.download_animation.currentPixmap()))
-
-    def stop_button_animation(self) -> None:
-        """
-        Stops the download animation and restores the button to its default state.
-        """
-        self.download_animation.stop()
-        self.file = self.controller.get_file(self.file.uuid)
-        self._set_file_state()
-
-
-class ModalDialog(QDialog):
-
-    BUTTON_CSS = load_css("modal_dialog_button.css")
-    ERROR_DETAILS_CSS = load_css("modal_dialog_error_details.css")
-
-    MARGIN = 40
-    NO_MARGIN = 0
-
-    def __init__(self, show_header: bool = True, dangerous: bool = False) -> None:
-        parent = QApplication.activeWindow()
-        super().__init__(parent)
-        self.setObjectName("ModalDialog")
-        self.setModal(True)
-
-        self.show_header = show_header
-        self.dangerous = dangerous
-        if self.dangerous:
-            self.setProperty("class", "dangerous")
-
-        # Widget for displaying error messages
-        self.error_details = QLabel()
-        self.error_details.setObjectName("ModalDialog_error_details")
-        self.error_details.setStyleSheet(self.ERROR_DETAILS_CSS)
-        self.error_details.setWordWrap(True)
-        self.error_details.hide()
-
-        # Body to display instructions and forms
-        self.body = QLabel()
-        self.body.setObjectName("ModalDialog_body")
-        self.body.setWordWrap(True)
-        self.body.setScaledContents(True)
-        body_container = QWidget()
-        self.body_layout = QVBoxLayout()
-        self.body_layout.setContentsMargins(
-            self.NO_MARGIN, self.NO_MARGIN, self.NO_MARGIN, self.NO_MARGIN
-        )
-        body_container.setLayout(self.body_layout)
-        self.body_layout.addWidget(self.body)
-
-        # Main widget layout
-        layout = QVBoxLayout(self)
-        layout.setContentsMargins(self.MARGIN, self.MARGIN, self.MARGIN, self.MARGIN)
-        self.setLayout(layout)
-
-        if self.show_header:
-            # Header for icon and task title
-            header_container = QWidget()
-            header_container_layout = QHBoxLayout()
-            header_container.setLayout(header_container_layout)
-            self.header_icon = SvgLabel("blank.svg", svg_size=QSize(64, 64))
-            self.header_icon.setObjectName("ModalDialog_header_icon")
-            self.header_spinner = QPixmap()
-            self.header_spinner_label = QLabel()
-            self.header_spinner_label.setObjectName("ModalDialog_header_spinner")
-            self.header_spinner_label.setMinimumSize(64, 64)
-            self.header_spinner_label.setVisible(False)
-            self.header_spinner_label.setPixmap(self.header_spinner)
-            self.header = QLabel()
-            self.header.setObjectName("ModalDialog_header")
-            header_container_layout.addWidget(self.header_icon)
-            header_container_layout.addWidget(self.header_spinner_label)
-            header_container_layout.addWidget(self.header, alignment=Qt.AlignCenter)
-            header_container_layout.addStretch()
-
-            self.header_line = QWidget()
-            self.header_line.setObjectName("ModalDialog_header_line")
-
-            layout.addWidget(header_container)
-            layout.addWidget(self.header_line)
-
-        layout.addWidget(self.error_details)
-        layout.addWidget(body_container)
-        layout.addWidget(self.configure_buttons())
-
-        # Activestate animation.
-        self.button_animation = load_movie("activestate-wide.gif")
-        self.button_animation.setScaledSize(QSize(32, 32))
-        self.button_animation.frameChanged.connect(self.animate_activestate)
-
-        # Header animation.
-        self.header_animation = load_movie("header_animation.gif")
-        self.header_animation.setScaledSize(QSize(64, 64))
-        self.header_animation.frameChanged.connect(self.animate_header)
-
-    def configure_buttons(self) -> QWidget:
-        # Buttons to continue and cancel
-        window_buttons = QWidget()
-        window_buttons.setObjectName("ModalDialog_window_buttons")
-
-        button_layout = QVBoxLayout()
-        window_buttons.setLayout(button_layout)
-
-        self.cancel_button = QPushButton(_("CANCEL"))
-        self.cancel_button.setStyleSheet(self.BUTTON_CSS)
-        self.cancel_button.clicked.connect(self.close)
-
-        self.continue_button = QPushButton(_("CONTINUE"))
-        self.continue_button.setStyleSheet(self.BUTTON_CSS)
-        self.continue_button.setIconSize(QSize(21, 21))
-
-        button_box = QDialogButtonBox(Qt.Horizontal)
-        button_box.setObjectName("ModalDialog_button_box")
-
-        if self.dangerous:
-            self.cancel_button.setAutoDefault(True)
-            self.continue_button.setDefault(False)
-            self.cancel_button.setObjectName("ModalDialog_primary_button")
-            self.continue_button.setObjectName("ModalDialog_cancel_button")
-        else:
-            self.cancel_button.setAutoDefault(False)
-            self.continue_button.setDefault(True)
-            self.cancel_button.setObjectName("ModalDialog_cancel_button")
-            self.continue_button.setObjectName("ModalDialog_primary_button")
-
-        button_box.addButton(self.cancel_button, QDialogButtonBox.ActionRole)
-        button_box.addButton(self.continue_button, QDialogButtonBox.ActionRole)
-
-        self.confirmation_label = QLabel()
-        self.confirmation_label.setObjectName("ModalDialogConfirmation")
-        button_layout.addWidget(self.confirmation_label, 0, Qt.AlignLeft | Qt.AlignBottom)
-
-        button_layout.addWidget(button_box, alignment=Qt.AlignLeft)
-
-        button_layout.setContentsMargins(
-            self.NO_MARGIN, self.NO_MARGIN, self.NO_MARGIN, self.NO_MARGIN
-        )
-
-        return window_buttons
-
-    def keyPressEvent(self, event: QKeyEvent) -> None:
-        if event.key() == Qt.Key_Enter or event.key() == Qt.Key_Return:
-            if self.cancel_button.hasFocus():
-                self.cancel_button.click()
-            else:
-                self.continue_button.click()
-        else:
-            super().keyPressEvent(event)
-
-    def animate_activestate(self) -> None:
-        self.continue_button.setIcon(QIcon(self.button_animation.currentPixmap()))
-
-    def animate_header(self) -> None:
-        self.header_spinner_label.setPixmap(self.header_animation.currentPixmap())
-
-    def start_animate_activestate(self) -> None:
-        self.button_animation.start()
-        self.continue_button.setText("")
-        self.continue_button.setMinimumSize(QSize(142, 43))
-        # Reset widget stylesheets
-        self.continue_button.setStyleSheet("")
-        self.continue_button.setObjectName("ModalDialog_primary_button_active")
-        self.continue_button.setStyleSheet(self.BUTTON_CSS)
-        self.error_details.setStyleSheet("")
-        self.error_details.setObjectName("ModalDialog_error_details_active")
-        self.error_details.setStyleSheet(self.ERROR_DETAILS_CSS)
-
-    def start_animate_header(self) -> None:
-        self.header_icon.setVisible(False)
-        self.header_spinner_label.setVisible(True)
-        self.header_animation.start()
-
-    def stop_animate_activestate(self) -> None:
-        self.continue_button.setIcon(QIcon())
-        self.button_animation.stop()
-        self.continue_button.setText(_("CONTINUE"))
-        # Reset widget stylesheets
-        self.continue_button.setStyleSheet("")
-        self.continue_button.setObjectName("ModalDialog_primary_button")
-        self.continue_button.setStyleSheet(self.BUTTON_CSS)
-        self.error_details.setStyleSheet("")
-        self.error_details.setObjectName("ModalDialog_error_details")
-        self.error_details.setStyleSheet(self.ERROR_DETAILS_CSS)
-
-    def stop_animate_header(self) -> None:
-        self.header_icon.setVisible(True)
-        self.header_spinner_label.setVisible(False)
-        self.header_animation.stop()
-
-
-class PrintDialog(ModalDialog):
-
-    FILENAME_WIDTH_PX = 260
-
-    def __init__(self, controller: Controller, file_uuid: str, file_name: str) -> None:
-        super().__init__()
-
-        self.controller = controller
-        self.file_uuid = file_uuid
-        self.file_name = SecureQLabel(
-            file_name, wordwrap=False, max_length=self.FILENAME_WIDTH_PX
-        ).text()
-        self.error_status = ""  # Hold onto the error status we receive from the Export VM
-
-        # Connect controller signals to slots
-        self.controller.export.printer_preflight_success.connect(self._on_preflight_success)
-        self.controller.export.printer_preflight_failure.connect(self._on_preflight_failure)
-
-        # Connect parent signals to slots
-        self.continue_button.setEnabled(False)
-        self.continue_button.clicked.connect(self._run_preflight)
-
-        # Dialog content
-        self.starting_header = _(
-            "Preparing to print:"
-            "<br />"
-            '<span style="font-weight:normal">{}</span>'.format(self.file_name)
-        )
-        self.ready_header = _(
-            "Ready to print:"
-            "<br />"
-            '<span style="font-weight:normal">{}</span>'.format(self.file_name)
-        )
-        self.insert_usb_header = _("Connect USB printer")
-        self.error_header = _("Printing failed")
-        self.starting_message = _(
-            "<h2>Managing printout risks</h2>"
-            "<b>QR codes and web addresses</b>"
-            "<br />"
-            "Never type in and open web addresses or scan QR codes contained in printed "
-            "documents without taking security precautions. If you are unsure how to "
-            "manage this risk, please contact your administrator."
-            "<br /><br />"
-            "<b>Printer dots</b>"
-            "<br />"
-            "Any part of a printed page may contain identifying information "
-            "invisible to the naked eye, such as printer dots. Please carefully "
-            "consider this risk when working with or publishing scanned printouts."
-        )
-        self.insert_usb_message = _("Please connect your printer to a USB port.")
-        self.generic_error_message = _("See your administrator for help.")
-
-        self._show_starting_instructions()
-        self.start_animate_header()
-        self._run_preflight()
-
-    def _show_starting_instructions(self) -> None:
-        self.header.setText(self.starting_header)
-        self.body.setText(self.starting_message)
-        self.error_details.hide()
-        self.adjustSize()
-
-    def _show_insert_usb_message(self) -> None:
-        self.continue_button.clicked.disconnect()
-        self.continue_button.clicked.connect(self._run_preflight)
-        self.header.setText(self.insert_usb_header)
-        self.body.setText(self.insert_usb_message)
-        self.error_details.hide()
-        self.adjustSize()
-
-    def _show_generic_error_message(self) -> None:
-        self.continue_button.clicked.disconnect()
-        self.continue_button.clicked.connect(self.close)
-        self.continue_button.setText(_("DONE"))
-        self.header.setText(self.error_header)
-        self.body.setText(  # nosemgrep: semgrep.untranslated-gui-string
-            "{}: {}".format(self.error_status, self.generic_error_message)
-        )
-        self.error_details.hide()
-        self.adjustSize()
-
-    @pyqtSlot()
-    def _run_preflight(self) -> None:
-        self.controller.run_printer_preflight_checks()
-
-    @pyqtSlot()
-    def _print_file(self) -> None:
-        self.controller.print_file(self.file_uuid)
-        self.close()
-
-    @pyqtSlot()
-    def _on_preflight_success(self) -> None:
-        # If the continue button is disabled then this is the result of a background preflight check
-        self.stop_animate_header()
-        self.header_icon.update_image("printer.svg", svg_size=QSize(64, 64))
-        self.header.setText(self.ready_header)
-        if not self.continue_button.isEnabled():
-            self.continue_button.clicked.disconnect()
-            self.continue_button.clicked.connect(self._print_file)
-            self.continue_button.setEnabled(True)
-            self.continue_button.setFocus()
-            return
-
-        self._print_file()
-
-    @pyqtSlot(object)
-    def _on_preflight_failure(self, error: ExportError) -> None:
-        self.stop_animate_header()
-        self.header_icon.update_image("printer.svg", svg_size=QSize(64, 64))
-        self.error_status = error.status
-        # If the continue button is disabled then this is the result of a background preflight check
-        if not self.continue_button.isEnabled():
-            self.continue_button.clicked.disconnect()
-            if error.status == ExportStatus.PRINTER_NOT_FOUND.value:
-                self.continue_button.clicked.connect(self._show_insert_usb_message)
-            else:
-                self.continue_button.clicked.connect(self._show_generic_error_message)
-
-            self.continue_button.setEnabled(True)
-            self.continue_button.setFocus()
-        else:
-            if error.status == ExportStatus.PRINTER_NOT_FOUND.value:
-                self._show_insert_usb_message()
-            else:
-                self._show_generic_error_message()
-
-
-class ExportDialog(ModalDialog):
-
-    PASSPHRASE_LABEL_SPACING = 0.5
-    NO_MARGIN = 0
-    FILENAME_WIDTH_PX = 260
-
-    def __init__(self, controller: Controller, file_uuid: str, file_name: str) -> None:
-        super().__init__()
-
-        self.controller = controller
-        self.file_uuid = file_uuid
-        self.file_name = SecureQLabel(
-            file_name, wordwrap=False, max_length=self.FILENAME_WIDTH_PX
-        ).text()
-        self.error_status = ""  # Hold onto the error status we receive from the Export VM
-
-        # Connect controller signals to slots
-        self.controller.export.preflight_check_call_success.connect(self._on_preflight_success)
-        self.controller.export.preflight_check_call_failure.connect(self._on_preflight_failure)
-        self.controller.export.export_usb_call_success.connect(self._on_export_success)
-        self.controller.export.export_usb_call_failure.connect(self._on_export_failure)
-
-        # Connect parent signals to slots
-        self.continue_button.setEnabled(False)
-        self.continue_button.clicked.connect(self._run_preflight)
-
-        # Dialog content
-        self.starting_header = _(
-            "Preparing to export:"
-            "<br />"
-            '<span style="font-weight:normal">{}</span>'.format(self.file_name)
-        )
-        self.ready_header = _(
-            "Ready to export:"
-            "<br />"
-            '<span style="font-weight:normal">{}</span>'.format(self.file_name)
-        )
-        self.insert_usb_header = _("Insert encrypted USB drive")
-        self.passphrase_header = _("Enter passphrase for USB drive")
-        self.success_header = _("Export successful")
-        self.error_header = _("Export failed")
-        self.starting_message = _(
-            "<h2>Understand the risks before exporting files</h2>"
-            "<b>Malware</b>"
-            "<br />"
-            "This workstation lets you open files securely. If you open files on another "
-            "computer, any embedded malware may spread to your computer or network. If you are "
-            "unsure how to manage this risk, please print the file, or contact your "
-            "administrator."
-            "<br /><br />"
-            "<b>Anonymity</b>"
-            "<br />"
-            "Files submitted by sources may contain information or hidden metadata that "
-            "identifies who they are. To protect your sources, please consider redacting files "
-            "before working with them on network-connected computers."
-        )
-        self.exporting_message = _("Exporting: {}").format(self.file_name)
-        self.insert_usb_message = _(
-            "Please insert one of the export drives provisioned specifically "
-            "for the SecureDrop Workstation."
-        )
-        self.usb_error_message = _(
-            "Either the drive is not encrypted or there is something else wrong with it."
-        )
-        self.passphrase_error_message = _("The passphrase provided did not work. Please try again.")
-        self.generic_error_message = _("See your administrator for help.")
-        self.continue_disabled_message = _(
-            "The CONTINUE button will be disabled until the Export VM is ready"
-        )
-        self.success_message = _(
-            "Remember to be careful when working with files outside of your Workstation machine."
-        )
-
-        # Passphrase Form
-        self.passphrase_form = QWidget()
-        self.passphrase_form.setObjectName("ExportDialog_passphrase_form")
-        passphrase_form_layout = QVBoxLayout()
-        passphrase_form_layout.setContentsMargins(
-            self.NO_MARGIN, self.NO_MARGIN, self.NO_MARGIN, self.NO_MARGIN
-        )
-        self.passphrase_form.setLayout(passphrase_form_layout)
-        passphrase_label = SecureQLabel(_("Passphrase"))
-        font = QFont()
-        font.setLetterSpacing(QFont.AbsoluteSpacing, self.PASSPHRASE_LABEL_SPACING)
-        passphrase_label.setFont(font)
-        self.passphrase_field = PasswordEdit(self)
-        self.passphrase_field.setEchoMode(QLineEdit.Password)
-        effect = QGraphicsDropShadowEffect(self)
-        effect.setOffset(0, -1)
-        effect.setBlurRadius(4)
-        effect.setColor(QColor("#aaa"))
-        self.passphrase_field.setGraphicsEffect(effect)
-        passphrase_form_layout.addWidget(passphrase_label)
-        passphrase_form_layout.addWidget(self.passphrase_field)
-        self.body_layout.addWidget(self.passphrase_form)
-        self.passphrase_form.hide()
-
-        self._show_starting_instructions()
-        self.start_animate_header()
-        self._run_preflight()
-
-    def _show_starting_instructions(self) -> None:
-        self.header.setText(self.starting_header)
-        self.body.setText(self.starting_message)
-        self.adjustSize()
-
-    def _show_passphrase_request_message(self) -> None:
-        self.continue_button.clicked.disconnect()
-        self.continue_button.clicked.connect(self._export_file)
-        self.header.setText(self.passphrase_header)
-        self.continue_button.setText(_("SUBMIT"))
-        self.header_line.hide()
-        self.error_details.hide()
-        self.body.hide()
-        self.passphrase_field.setFocus()
-        self.passphrase_form.show()
-        self.adjustSize()
-
-    def _show_passphrase_request_message_again(self) -> None:
-        self.continue_button.clicked.disconnect()
-        self.continue_button.clicked.connect(self._export_file)
-        self.header.setText(self.passphrase_header)
-        self.error_details.setText(self.passphrase_error_message)
-        self.continue_button.setText(_("SUBMIT"))
-        self.header_line.hide()
-        self.body.hide()
-        self.error_details.show()
-        self.passphrase_field.setFocus()
-        self.passphrase_form.show()
-        self.adjustSize()
-
-    def _show_success_message(self) -> None:
-        self.continue_button.clicked.disconnect()
-        self.continue_button.clicked.connect(self.close)
-        self.header.setText(self.success_header)
-        self.continue_button.setText(_("DONE"))
-        self.body.setText(self.success_message)
-        self.cancel_button.hide()
-        self.error_details.hide()
-        self.passphrase_form.hide()
-        self.header_line.show()
-        self.body.show()
-        self.adjustSize()
-
-    def _show_insert_usb_message(self) -> None:
-        self.continue_button.clicked.disconnect()
-        self.continue_button.clicked.connect(self._run_preflight)
-        self.header.setText(self.insert_usb_header)
-        self.continue_button.setText(_("CONTINUE"))
-        self.body.setText(self.insert_usb_message)
-        self.error_details.hide()
-        self.passphrase_form.hide()
-        self.header_line.show()
-        self.body.show()
-        self.adjustSize()
-
-    def _show_insert_encrypted_usb_message(self) -> None:
-        self.continue_button.clicked.disconnect()
-        self.continue_button.clicked.connect(self._run_preflight)
-        self.header.setText(self.insert_usb_header)
-        self.error_details.setText(self.usb_error_message)
-        self.continue_button.setText(_("CONTINUE"))
-        self.body.setText(self.insert_usb_message)
-        self.passphrase_form.hide()
-        self.header_line.show()
-        self.error_details.show()
-        self.body.show()
-        self.adjustSize()
-
-    def _show_generic_error_message(self) -> None:
-        self.continue_button.clicked.disconnect()
-        self.continue_button.clicked.connect(self.close)
-        self.continue_button.setText(_("DONE"))
-        self.header.setText(self.error_header)
-        self.body.setText(  # nosemgrep: semgrep.untranslated-gui-string
-            "{}: {}".format(self.error_status, self.generic_error_message)
-        )
-        self.error_details.hide()
-        self.passphrase_form.hide()
-        self.header_line.show()
-        self.body.show()
-        self.adjustSize()
-
-    @pyqtSlot()
-    def _run_preflight(self) -> None:
-        self.controller.run_export_preflight_checks()
-
-    @pyqtSlot()
-    def _export_file(self, checked: bool = False) -> None:
-        self.start_animate_activestate()
-        self.cancel_button.setEnabled(False)
-        self.passphrase_field.setDisabled(True)
-        self.controller.export_file_to_usb_drive(self.file_uuid, self.passphrase_field.text())
-
-    @pyqtSlot()
-    def _on_preflight_success(self) -> None:
-        # If the continue button is disabled then this is the result of a background preflight check
-        self.stop_animate_header()
-        self.header_icon.update_image("savetodisk.svg", QSize(64, 64))
-        self.header.setText(self.ready_header)
-        if not self.continue_button.isEnabled():
-            self.continue_button.clicked.disconnect()
-            self.continue_button.clicked.connect(self._show_passphrase_request_message)
-            self.continue_button.setEnabled(True)
-            self.continue_button.setFocus()
-            return
-
-        self._show_passphrase_request_message()
-
-    @pyqtSlot(object)
-    def _on_preflight_failure(self, error: ExportError) -> None:
-        self.stop_animate_header()
-        self.header_icon.update_image("savetodisk.svg", QSize(64, 64))
-        self._update_dialog(error.status)
-
-    @pyqtSlot()
-    def _on_export_success(self) -> None:
-        self.stop_animate_activestate()
-        self._show_success_message()
-
-    @pyqtSlot(object)
-    def _on_export_failure(self, error: ExportError) -> None:
-        self.stop_animate_activestate()
-        self.cancel_button.setEnabled(True)
-        self.passphrase_field.setDisabled(False)
-        self._update_dialog(error.status)
-
-    def _update_dialog(self, error_status: str) -> None:
-        self.error_status = error_status
-        # If the continue button is disabled then this is the result of a background preflight check
-        if not self.continue_button.isEnabled():
-            self.continue_button.clicked.disconnect()
-            if self.error_status == ExportStatus.BAD_PASSPHRASE.value:
-                self.continue_button.clicked.connect(self._show_passphrase_request_message_again)
-            elif self.error_status == ExportStatus.USB_NOT_CONNECTED.value:
-                self.continue_button.clicked.connect(self._show_insert_usb_message)
-            elif self.error_status == ExportStatus.DISK_ENCRYPTION_NOT_SUPPORTED_ERROR.value:
-                self.continue_button.clicked.connect(self._show_insert_encrypted_usb_message)
-            else:
-                self.continue_button.clicked.connect(self._show_generic_error_message)
-
-            self.continue_button.setEnabled(True)
-            self.continue_button.setFocus()
-        else:
-            if self.error_status == ExportStatus.BAD_PASSPHRASE.value:
-                self._show_passphrase_request_message_again()
-            elif self.error_status == ExportStatus.USB_NOT_CONNECTED.value:
-                self._show_insert_usb_message()
-            elif self.error_status == ExportStatus.DISK_ENCRYPTION_NOT_SUPPORTED_ERROR.value:
-                self._show_insert_encrypted_usb_message()
-            else:
-                self._show_generic_error_message()
 
 
 class DeleteSourceDialog(ModalDialog):
